Register detail is now stored on memory.
Register detail stored on memory until the user chooses a server.
Should be safe, since its a desktop app any malicious
hacker would have to have access to the computer to get the details.
And they're only stored for a short time until the user chooses a server.
or refreshes the page. or exits.

Made the buttons look nice with icons and everything is working smoothly.
No issues so far.