Added animation from the signup to the two buttons
To the "add server and search for domain" buttons.
Also switched up the animation a little bit, not sure exactly
how but now both the initial log in to sign up animation and
sign up to buttons animation is connected.

No issues yet. Other than the two buttons not following their css properly.
Will fix that in the next commit.