Added working server functionality to my code.
You can now add a MySQL, PostgreSQL and SQLite servers.
I tested it using docker and it worked with MySQL, the others should
work as expected. I also made the form for the button open and close
when pressed again.

In the next commit I will make sure the buttons have nice
animations and that the css looks nice and clean.

No apparent issues so far.