The register detail is now saved inside database.
I had already implemented server functionality, all that
was needed to be done was store the register details inside
the database. This code does exactly that, it creates a table
inside the database called vantablack-users and adds the register
details there. No way to login yet, that comes in the next commit.

I also removed the storage of server-details in local storage
using JSON stringify since it was useless.

No apparent issues. This code is good to merge with develop.